Fix invalid XML
Also, not quite sure what you're trying to do with the extra volume.  If it's a mistake, then you need to delete that volume you added.  If it's supposed to be there, then it needs to be another <Volume> section, not within the existing one, and you would also need a new <Config> for the added path.  How you've done it will have the feed trigger a fatal error, since it's not correct.